[2/3] Split Signature into Signature and Method classes (preparatory work for pretty PR) (#112)
This PR does not contain any of the 'pretty printing code' (which is messy and long), but contains only the internal changes that make it possible to implement.

Mainly, this PR removes `ìmplementaition` and `return type` from `Signature` and introduces a new class , `Method` that pieces those things together (a method is just a signature, an implementation, and a name).

Conceptually I think this makes sense, because the signatures should only be used to dispatch. 
Also, I think this might make some experimentation easier later on... 

Apart from this two changes, everything is pretty mechanical. @wesselb there is one failing test that I don't understand why it fails. Do you have an idea?